+++ src/main/java/com/epam/ta/reportportal/ws/model/ErrorType.java
@@ -3,7 +3,7 @@
  * 
  * 
  * This file is part of EPAM Report Portal.
- * https://github.com/epam/ReportPortal
+ * https://github.com/reportportal/commons-model
  * 
  * Report Portal is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -371,6 +371,11 @@ public enum ErrorType {
 	 * Incorrect update preference request
 	 */
 	BAD_UPDATE_PREFERENCE_REQUEST(40032, "Incorrect update widget request {}"),
+
+	/**
+	 * Unsupported merge strategy type
+	 */
+	UNSUPPORTED_MERGE_STRATEGY_TYPE(40033, "Merge Strategy type {} is unsupported"),
 
 	/**
 	 * Use it If there are no any other exceptions. There should by no such